pyproject.toml
[tool.poetry]
name = "foreflow"
version = "0.1.0"
description = ""
authors = ["Naoya Yamashita <conao3@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.scripts]
foreflow = "foreflow.cli:main"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.2"
pyyaml = "^6.0.1"
jmespath = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-icdiff = "^0.8"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
